instrumentation/python/traces/redis/app/requirements.txt
flask
#opentelemetry-exporter-otlp
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask>=0.8b0
opentelemetry-instrumentation-requests>=0.8b0
#opentelemetry-instrumentation-sqlite3>=0.11b0
opentelemetry-instrumentation-redis>=0.37b0
requests
redis